projetos/rotatividade-de-clientes/machine-learning/requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
imblearn
xgboost
pyyaml
joblib
